Add Wallhaven.API subsystem
