+++ src/Types.hs
@@ -1,118 +1,8 @@
 module Types where
 
 import Control.Monad.Reader (MonadReader, asks, liftIO)
-import Retry (HasRetryConfig, RetryConfig, getRetryConfig)
 import UnliftIO (MonadIO, hFlush, stdout)
 import Prelude hiding (log)
-
-data Env = Env
-  { envConfig :: !Config,
-    envLog :: !(String -> IO ())
-  }
-
--- Configuration structure.
-data Config = Config
-  { -- | The directory where the wallpapers will be saved.
-    configWallpaperDir :: FilePath,
-    -- | The number of wallpapers to download in parallel.
-    configNumParallelDownloads :: NumParallelDownloads,
-    -- | Retry config for HTTP requests.
-    configRetryConfig :: RetryConfig,
-    -- | Whether to delete unliked wallpapers.
-    configDeleteUnliked :: Bool,
-    -- | Wallhaven usename
-    configUsername :: String,
-    -- | Wallhaven API Key
-    configWallhavenAPIKey :: String,
-    -- | Label of the collection to sync
-    configCollectionLabel :: String,
-    -- | Debug mode
-    configDebug :: Bool
-  }
-
-class HasDebug a where
-  getDebug :: a -> Bool
-
-instance HasDebug Config where
-  getDebug = configDebug
-
-instance HasDebug Env where
-  getDebug = getDebug . envConfig
-
-class HasCollectionLabel a where
-  getCollectionLabel :: a -> String
-
-instance HasCollectionLabel Config where
-  getCollectionLabel = configCollectionLabel
-
-instance HasCollectionLabel Env where
-  getCollectionLabel = getCollectionLabel . envConfig
-
-class HasWallhavenUsername a where
-  getWallhavenUsername :: a -> String
-
-instance HasWallhavenUsername Config where
-  getWallhavenUsername = configUsername
-
-instance HasWallhavenUsername Env where
-  getWallhavenUsername = getWallhavenUsername . envConfig
-
-class HasWallhavenAPIKey a where
-  getWallhavenAPIKey :: a -> String
-
-instance HasWallhavenAPIKey Config where
-  getWallhavenAPIKey = configWallhavenAPIKey
-
-instance HasWallhavenAPIKey Env where
-  getWallhavenAPIKey = getWallhavenAPIKey . envConfig
-
-class HasDeleteUnliked a where
-  getDeleteUnliked :: a -> Bool
-
-instance HasDeleteUnliked Config where
-  getDeleteUnliked = configDeleteUnliked
-
-instance HasDeleteUnliked Env where
-  getDeleteUnliked = getDeleteUnliked . envConfig
-
-class HasLog a where
-  getLog :: a -> (String -> IO ())
-
-instance HasLog Env where
-  getLog = envLog
-
-class HasNumParallelDownloads a where
-  getNumParallelDownloads :: a -> NumParallelDownloads
-
-instance HasNumParallelDownloads Config where
-  getNumParallelDownloads = configNumParallelDownloads
-
-instance HasNumParallelDownloads Env where
-  getNumParallelDownloads = getNumParallelDownloads . envConfig
-
-class HasWallpaperDir a where
-  getWallpaperDir :: a -> FilePath
-
-instance HasWallpaperDir Config where
-  getWallpaperDir = configWallpaperDir
-
-instance HasWallpaperDir Env where
-  getWallpaperDir = getWallpaperDir . envConfig
-
-instance HasRetryConfig Config where
-  getRetryConfig = configRetryConfig
-
-instance HasRetryConfig Env where
-  getRetryConfig = getRetryConfig . envConfig
-
-class HasConfig a where
-  getConfig :: a -> Config
-
-instance HasConfig Config where
-  getConfig = id
-
-instance HasConfig Env where
-  getConfig = envConfig
 
 type FullWallpaperURL = String
 
@@ -132,14 +22,3 @@ type CollectionID = Int
 type WallpaperPath = String
 
 type Username = String
-
-type APIKey = String
-
-log :: (MonadReader r m, HasLog r, MonadIO m) => String -> m ()
-log !msg = do
-  logger <- asks getLog
-  liftIO $ logger msg
-  hFlush stdout
-
-logLn :: (MonadReader r m, HasLog r, MonadIO m) => String -> m ()
-logLn msg = log (msg <> "\n")
