+++ src/Wallhaven/API/Class.hs
@@ -1,0 +1,10 @@
+module Wallhaven.API.Class where
+
+import Types (NumParallelDownloads)
+import Wallhaven.API.Types (APIKey)
+
+class HasWallhavenAPIKey a where
+  getWallhavenAPIKey :: a -> APIKey
+
+class HasNumParallelDownloads a where
+  getNumParallelDownloads :: a -> NumParallelDownloads
